+++ CHANGES.md
@@ -21,7 +21,7 @@ Changes:
 -
 
 Fixes:
--
+- Fix `nd` scraper crashing on rows without a 'View Opinion' button #1872
 
 ## 3.0.2 - 2026-03-20